Pick a Contact app page split code into files
