+++ imessage_extractor/src/quality_control/views/qc_message_special_types.sql
@@ -1,7 +1,7 @@
 drop view if exists {pg_schema}.qc_message_special_types;
 create or replace view {pg_schema}.qc_message_special_types as
 
-select associated_type, message_special_type, count(*) as n
+select associated_type, message_special_type, count(message_id) as n_messages
 from {pg_schema}.message_vw
 where message_special_type is null
   and associated_type != 0
